refactor: 라우터 경로 수정
auth.js 경로 수정

회원가입 기본 특성 선택

회원가입 카드 위치 수정
